postgres_to_es/requirements/base.txt
certifi==2021.10.8
elasticsearch==7.15.1
elasticsearch-dsl==7.4.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
requests==2.26.0

